aep/orderingtable: call the real neGraphics API (fix the link error)
AepOrderingTable.mm forward-declared the neGraphics primitives in a local
extern "C" block with stale signatures, so the C++ definitions never resolved
and the build failed at link with 7 undefined symbols
(neDrawLine/Triangle/Rect/Quad/TexturedQuad/Text + setRenderStateSlot).

Drop the local block and use the real declarations:
- #import neRenderer.h / neTextTexture.h / neTextureRef.h.
- drawAepOtText now calls neDrawText(text, font, size, x, y, align, alpha, r, g,
  b, clipRect) (FUN_0001551c); the binary passes p3 as the font handle and
  colorVec as the clip rect.
- drawAepSpriteClipped resolves the per-frame neTextureRef record at
  [frameObj+0x14] + frame*0x18 (Ghidra 0x12154) and uses its 2-arg
  setRenderStateSlot method (0x1215e/0x12180); that same record is the sprite
  object passed to neDrawTexturedQuad (0x121f8 -> FUN_00015fb8), whose argument
  values line up positionally with the real 18-parameter signature (the trailing
  p16 flag is not one of them).
- neDrawTexturedQuad's clipRect is float, not int (drawAepSpriteClipped and the
  binary build a float[4] clip via vcvt.f32.s32), so retype it const float * in
  neRenderer.h/.cpp to match.

+++ Project/System/src/Render/neRenderer.cpp
@@ -404,7 +404,7 @@ void neDrawTexturedQuad(void *sprite,
                         int green,
                         int blue,
                         int blendMode,
-                        const int *clipRect) {
+                        const float *clipRect) {
     neSpriteView *s = static_cast<neSpriteView *>(sprite);
     uint32_t c = nePremultRGBA(alpha, red, green, blue);
